Moved version extraction logic to separate file.
